old_structure/FINAL_TOOLS_OUTPUT/requirements.txt
pandas==2.2.0
selenium==4.15.2
undetected-chromedriver==3.5.4
beautifulsoup4==4.12.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
